requirements.txt
google-generativeai
flask
httpx
python-dotenv
twilio